Tar høyde for minimal person og sjekker om det er personinfo som skal slettes
Hvis delete-query opprettes når personinfoRef er en tom liste vil det bli en syntaksfeil ved kjøring.

+++ spesialist-opprydding-dev/src/main/kotlin/no/nav/helse/opprydding/PersonRepository.kt
@@ -63,6 +63,7 @@ internal class PersonRepository(private val dataSource: DataSource) {
     }
 
     private fun TransactionalSession.slettInfo(personinfoRef: List<Int>) {
+        if (personinfoRef.isEmpty()) return
         @Language("PostgreSQL")
         val query = "DELETE FROM person_info WHERE id IN (${personinfoRef.joinToString { "?" }})"
         run(queryOf(query, *personinfoRef.toTypedArray()).asExecute)
